NTEventLogAppender: Improved internal consistency; deprecated open()
The old open() function was only safe to call when the appender is in closed state, otherwise it would leak a handle.
Also, the open() function returned void, so when it failed, it failed silently.

Therefore it made sense to move the functionality of the open() function to the reopen() function, so that close() is called first, and so that any failure can be reported as a "false" result.

The open() function now delegates to reopen(). I added a \deprecated tag so that Doxygen will flag it as such. I didn't add a __declspec(deprecated) because I think Visual C++ 6 won't handle it correctly.

I also rewrote the comments for reopen() and close() to indicate that these functions correctly handle corner cases (opening when already open, closing when already closed).