NTEventLogAppender: Improved flexibility, fixed bug 148
The old version of addRegistryInfo would always use "NTEventLogAppender.dll" without drive or directory, so it was necessary to put the DLL in the PATH or change the PATH to add the directory where the DLL is stored. This is undesirable and unnecessary: if the full path of the DLL is stored in the registry, it doesn' t need to be in the PATH.

Another problem was that the reopen() function would always get the registry key to check if it existed, which should not be necessary in most cases. And if the registry key already existed, the function wouldn't save the values so it was impossible to deliberately change the values.

Furthermore, the addRegistryInfo() function returned void, which made it impossible to notify the caller of a failure, which is not impossible; keep in mind that the registry can only be written if the program runs as administrator.

Finally, the protected registry functions used TCHAR in their prototypes, which is a very bad idea: programs that are compiled in Unicode and use these functions would be incompatible with the log4cpp.dll because the DLL and the program would disagree on what a TCHAR is. This results in linker problems that are potentially very difficult or impossible to debug, especially for beginning Windows programmers.

I moved the registry modifying code into an overloaded reopen() function that takes a parameter for the name, and a parameter for the DLL location. If the DLL location is blank, it uses "NTEventLogAppender.dll" for backwards compatibility.

reopen() with parameters always attempts to write the values into the registry unless it's unable to open the key.

If the source name (which is used as registry subkey) is blank, the parameterized reopen() fails.

The non-parameterized reopen() doesn't operate on the registry, it merely calls ::RegisterEventSource.

The protected functions that operate on the registry were corrected to use LPCSTR (const char *) instead of TCHAR * (without const). Existing software that calls these functions should still work.

Two constructors were added: one that accepts the DLL path as a parameter and one that creates the instance in closed state, i.e. without calling reopen(). This fixes the problem that derived classes might not want the constructor to call the reopen() function, and because virtual functions aren't really virtual when called from a constructor, it would not be enough to override the function in the derived class.

Other changes:
- I modified the class comment slightly to reflect that it's possible to provide a full path to the NTEventLogAppender.dll now.
- The protected addRegistryInfo function is now deprecated so it has been marked as such with a \deprecated tag. I didn't use __declspec(deprecated) because I think Visual C++ 6 won't understand it.