VS2010 build system fix: Add ProjectGUIDs
Added ProjectGUIDs to the test projects that didn't have one.

Visual Studio has a known bug where under some circumstances (e.g. when updating a project created in a previous version of Visual Studio) it doesn't store a project GUID in the project file. But it needs project GUIDs in the solution file to keep track of relations between projects, and if it doesn't find a GUID in the project file, it makes up a new one. So everytime you open the solution, it makes changes to it and wants you to check them into your source control system.

The solution is to manually add the project GUID to the project files which is what I did here.