VS2010 build system: Create a separate project for NTEventLogAppender.dll
In the old situation, the NTEventLogAppender.dll resource DLL was created with custom commands, and its compilation was duplicated in no less than THREE projects.

On at least one of my machines, this resulted in a problem because two separate build threads wanted to create the DLL at the same time, and one thread would fail because the other thread was writing the DLL.

I created a new project file so that Debug and Release versions of this resource DLL are only created once, and the build system won't attempt to overwrite a build that it just created as part of the same solution build.

I changed the custom build command for the .mc file slightly: I changed the -h and -r parameter so they use $(IntDir). That directory is excluded in .gitignore so that the newly created files don't accidentally get added to source control.

I removed the commands to compile the resource file from the custom build, since it's just as easy to add the .rc file to the project. Visual Studio can figure out how to compile the resource into the DLL. The only thing that was left to do was to make sure the .rc file in the release directory is not compiled in the debug build and vice versa; I did this by adding both files (debug and release) but excluding each file from the other build.

I removed the .mc file from the projects that previously had it.

I added a dependency on the NTEventLogAppender project to each of the three projects that formerly had the .mc file. Adding a project reference doesn't work because if you do that, Visual Studio goes looking for a library which doesn't exist because there's no code in the DLL.